feat: Update ONNX model and load MinMaxScaler
This commit updates the ONNX model file to 'best_lstm_yolo_model_20240902_1930.onnx' and updates the ONNX Runtime session to load the new model. Additionally, it adds code to load the saved MinMaxScaler from 'scaler_X.pkl' and 'scaler_y.pkl'. These changes are made to improve the accuracy and performance of the model.